Move get_search_results to the proper class
